refactor(desktop): decompose AgentChatThread into focused virtualization, scroll, and row-render modules
